server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
Arpeggio==2.0.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.8
dill==0.3.7
Django==4.2.2
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-environ==0.10.0
django-simple-deploy==0.5.15
djangorestframework==3.14.0
djlint==1.31.1
docopt==0.6.2
EditorConfig==0.12.3
frozenlist==1.4.0
grpcio==1.56.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.8
json5==0.9.14
markdown2==2.4.9
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.8
packaging==23.1
parver==0.4
pathspec==0.11.1
pipreqs==0.4.13
platformdirs==3.8.0
protobuf==4.23.4
pulumi==3.76.0
pulumi-gcp==6.60.0
pycodestyle==2.10.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
semver==2.13.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tqdm==4.65.0
typing_extensions @ file:///private/tmp/python-typing-extensions-20230601-7330-1g8neex/typing_extensions-4.6.3/dist/typing_extensions-4.6.3-py3-none-any.whl
urllib3==2.0.3
yarg==0.1.9
yarl==1.9.2
